requirements/dev.txt
-r prod.txt
black==24.8.0
django-debug-toolbar==4.3.0